Don't attempt to swap equipment sets when pausing
When fishing mode automatically pauses, it's because we are about to
enter combat lockdown. We are given the chance to adjust settings and
hide frames, but attempting to swap equipment fails and the game does
not report the failure to us. Since we know this will always fail, we
just avoid trying to change equipment in this transitional state.
Doing this prevents us from deleting the backup equipment set before
we've had a chance to actually restore the items onto the player.