Handle pausing/resuming better
Now that fishing mode can pause while out of combat, we have to handle
the case where we are paused and the player tries to toggle fishing
mode. This was unnecessary previously because we simply denied any
attempts to toggle fishing mode while in combat due to limitations of
the lockdown.

We have also introduced a new orange icon state to indicate that fishing
mode has been paused.